# back/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Autentication Users",
        "descriptions": "Serviço com foco em validações e gerenciamente de permissões de usuarios",
        "contact": {
            "email": "nivaldeir-silva@hotmail.com"
        }
    },
    "paths": {
...
        "/user": {
            "post": {
                "summary": "Cadastro de usuario",
                "descriptions": "",
                "tags": ["User"],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            },
                            "examples": {
                                "user": {
                                    "value": {
                                        "email": "test@test.com.br",
                                        "username": "test",
                                        "password": "test",
                                        "id": "optional"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Já existe informação com esse...",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "summary": "Busca de usuario",
                "descriptions": "",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": ["User"],
                "responses": {
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "$ref": "#/components/schemas/UserArray"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/{id}": {
            "get": {
                "descriptions": "Busca de usuario pelo ID",
                "summary": "Busca de usuario pelo ID",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": ["User"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do usuario",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "descriptions": "Atualizando um usuario pelo ID",
                "summary": "Atualizando um usuario pelo ID",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": ["User"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do usuario",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            },
                            "examples": {
                                "user": {
                                    "value": {
                                        "username": "optional",
                                        "password": "optional"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "descriptions": "Desativando um usuario pelo ID",
                "summary": "Desativando um usuario pelo ID",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": ["User"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do usuario",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/{id}/deleting-permission": {
            "put": {
                "descriptions": "Remoção permissão do usuario pelo ID",
                "summary": "Remoção permissão do usuario pelo ID",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": ["User"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do usuario",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "examples": {
                                "user": {
                                    "value": {
                                        "permission": "create_user"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "message": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/{id}/adding-permission": {
            "put": {
                "descriptions": "Adicionando permissão do usuario pelo ID",
                "summary": "Adicionando permissão do usuario pelo ID",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": ["User"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do usuario",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "examples": {
                                "user": {
                                    "value": {
                                        "permission": "create_user"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "message": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "UserArray": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "email": {
                                    "type": "string"
                                },
                                "enabled": {
                                    "type": "boolean"
                                },
                                "id": {
                                    "type": "string"
                                },
                                "roleId": {
                                    "type": "string"
                                },
                                "username": {
                                    "type": "string"
                                }
                            },
                            "required": [
                                "email",
                                "enabled",
                                "id",
                                "roleId",
                                "username"
                            ]
                        }
                    }
                },
                "required": ["message", "data"]
            },
            "User": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "object",
                        "properties": {
                            "email": {
                                "type": "string"
                            },
                            "enabled": {
                                "type": "boolean"
                            },
                            "id": {
                                "type": "string"
                            },
                            "roleId": {
                                "type": "string"
                            },
                            "username": {
                                "type": "string"
                            },
                            "password": {
                                "type": "string"
                            },
                            "salt": {
                                "type": "string"
                            }
                        }
                    }
                },
                "Role": {
                    "type": "object",
                    "properties": {
                        "name": {
                            "type": "string"
                        }
                    }
                },
                "Permission": {
                    "type": "object",
                    "properties": {
                        "name": {
                            "type": "string"
                        }
                    }
                },
                "Auth": {
                    "type": "object",
                    "properties": {
                        "password": {
                            "type": "string"
                        },
                        "email": {
                            "type": "string"
                        },
                        "username": {
                            "type": "string"
                        }
                    }
                },
                "ResponseError": {
                    "type": "object",
                    "properties": {
                        "error": {
                            "type": "boolean"
                        },
                        "message": {
                            "type": "string"
                        }
                    }
                }
            },
...
            "Error": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    }
                }
            },
...
        }
    }
}